pyproject.toml
[tool.poetry]
name = "curso-de-python-do-basico-ao-avancado"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "curso_de_python_do_basico_ao_avancado"}]

[tool.poetry.dependencies]
python = "^3.11"
webdriver-manager = "^3.8.5"
bs4 = "^0.0.1"
selenium = "^4.8.2"
requests = "^2.28.2"
types-requests = "^2.28.11.15"
python-dotenv = "^1.0.0"
types-python-dateutil = "^2.8.19.10"
python-dateutil = "^2.8.2"
types-pytz = "^2022.7.1.2"
pytz = "^2022.7.1"
notebook = "^6.5.3"
matplotlib = "^3.7.1"
numpy = "^1.24.2"
pandas = "^1.5.3"
pypdf2 = "^3.0.1"
openpyxl = "^3.1.2"
pillow = "^9.4.0"
PySide6 = "^6.4.2"

[tool.poetry.group.doc.dependencies]
mkdocs = "^1.4.2"
mkdocs-material = "^9.1.1"
mkdocstrings = "^0.20.0"
mkdocs-localsearch = "^0.9.2"
mkdocs-redirects = "^1.2.0"
mkdocs-minify-plugin = "^0.6.2"
mkdocs-glightbox = "^0.3.1"


[tool.poetry.group.dev.dependencies]
mkdocs-glightbox = "^0.3.1"
blue = "^0.9.1"
isort = "^5.12.0"
pytest-cov = "^4.0.0"
pytest = "^7.2.2"
taskipy = "^1.10.3"
mypy = "^1.1.1"
cairosvg = "^2.6.0"
pillow = "^9.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
line_length = 80

[tool.taskipy.tasks]
lint = "blue --check --diff . && isort --check --diff ."
docs = "mkdocs serve"
pre_test = "task lint"
post_test = "coverage html"